+++ packages/sources/react/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.22.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/react@0.22.0...@vtmn/react@0.22.1) (2022-03-07)
+
+
+### Bug Fixes
+
+* **@vtmn/react:** add missing className, {...props} props in `VtmnProgressbar` ([#1022](https://github.com/Decathlon/vitamin-web/issues/1022)) ([652932d](https://github.com/Decathlon/vitamin-web/commit/652932d7171fcd3aa13f3bfbc43b90bf6c7b0f4f))
+
+
+
+
+
+# [0.22.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/react@0.21.4...@vtmn/react@0.22.0) (2022-03-06)
+
+
+### Features
+
+* **@vtmn/react:** add `VtmnProgressbar` component ([#1018](https://github.com/Decathlon/vitamin-web/issues/1018)) ([f1b68cd](https://github.com/Decathlon/vitamin-web/commit/f1b68cd7eba5b24f001da615fb0bcf190ff44f17))
+
+
+
+
+
 ## [0.21.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/react@0.21.3...@vtmn/react@0.21.4) (2022-03-04)
 
 **Note:** Version bump only for package @vtmn/react